backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1
openai==1.59.7
python-dateutil==2.9.0.post0
